requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
APScheduler==3.11.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.7.0
geoip2==5.1.0
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Js2Py==0.74
Mako==1.3.10
MarkupSafe==3.0.2
maxminddb==2.8.2
multidict==6.6.4
packaging==25.0
pipwin==0.5.2
propcache==0.3.2
pycparser==2.22
pyjsparser==2.7.1
PyMySQL==1.1.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.42
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.20.1
